requirements.txt
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.33.25
backoff==2.2.1
boto3==1.34.123
botocore==1.34.143
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.3
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
inference==0.13.0
inference-sdk==0.13.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.15.1
openai==1.35.13
opencv-python==4.8.0.76
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pendulum==3.0.0
pexpect==4.9.0
piexif==1.1.3
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybase64==1.3.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydot==2.0.0
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.2
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.7
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
roboflow==1.1.34
rpds-py==0.19.0
rsa==4.7.2
s3transfer==0.10.2
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.4
six==1.16.0
skypilot==0.5.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
structlog==24.2.0
supervision==0.21.0
sympy==1.13.0
tabulate==0.9.0
tifffile==2024.7.2
time-machine==2.14.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.19
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.9.4
zxing-cpp==2.2.0
